Renamed some files to conform to eslint rule
