+++ app/containers/problems.js
@@ -3,12 +3,12 @@ import React from 'react';
 import styled, { keyframes } from 'styled-components';
 import { fadeIn } from 'react-animations';
 
-import Text from '../components/Text';
-import Error from '../components/Error';
-import Loading from '../components/Loading';
-import withSettings from './withSettings';
+import Text from '../components/text';
+import Error from '../components/error';
+import Loading from '../components/loading';
+import withSettings from './with-settings';
 import getFontSize from '../util/get-font-size';
-import { ScrollerBase } from '../components/ScrollerBase';
+import { ScrollerBase } from '../components/scroller-base';
 
 const fadeInAnimation = keyframes`${fadeIn}`;
