+++ internals/mocks/file-mock.js
