+++ internals/scripts/check-builts-exist.js
